+++ models/user.js
@@ -1,10 +1,27 @@
 'use strict';
 module.exports = (sequelize, DataTypes) => {
   var user = sequelize.define('user', {
-      id:{type:DataTypes.BIGINT, autoIncrement:true, primaryKey : true, unique : true},
-      username: { type: DataTypes.STRING,  allowNull: false, comment:'用户名' },
-      password: { type: DataTypes.STRING, allowNull: false, comment:'用户密码' },
-      active: { type: DataTypes.BOOLEAN, allowNull: false, defaultValue: true, comment:'是否正常状态' }
+      id:{
+          type: DataTypes.INTEGER,
+          autoIncrement:true,
+          primaryKey : true,
+          unique : true
+      },
+      username: {
+          type: DataTypes.STRING,
+          allowNull: false,
+          comment:'用户名'
+      },
+      password: {
+          type: DataTypes.STRING,
+          allowNull: false,
+          comment:'用户密码' },
+      active: {
+          type: DataTypes.BOOLEAN,
+          allowNull: false,
+          defaultValue: true,
+          comment:'是否正常状态'
+      }
   }, {
       timestamps: true,
       underscored: true,
